feat:add previewCard component
