+++ src/components/previewCard/index.module.css
@@ -1,0 +1,3 @@
+.promotionImage {
+    height: 209px;
+  }
